requirements.txt
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
httpie==3.2.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==70.1.0
six==1.16.0
urllib3==2.2.2
Werkzeug==3.0.3
